fix acceptance test skip_ssl_validation
previously filled this value in based on riak_cs.skip_ssl_validation.
that value is used to determine how the broker communicates with the
riak cs nodes.

the tests use this value to determine how the cli communicates with cf
runtime.  create a new setting for this field: cf.skip_ssl_validation